Fix parameter type mismatch in ListPromMetricTags query
The SQL query expected a scalar int64 for metric_fingerprint but sqlc
generated []int64, causing runtime errors when executing the query.

Fixed by:
- Change SQL parameter from []int64 to int64 with ::BIGINT cast
- Update handler to pass scalar fingerprint value
- Add integration tests to verify metric filtering works correctly

The tests confirm that:
- Tags are correctly filtered by metric fingerprint
- Only tags for the requested metric are returned
- Underscored column names are returned (not dotted)
- Empty results when no segments match

+++ lrdb/queries/promql_tags.sql
@@ -17,12 +17,12 @@ LIMIT 1;
 -- name: ListPromMetricTags :many
 -- Extract tag keys from label_name_map in metric_seg table for a specific metric
 -- Filters by metric fingerprint to return tags only for the requested metric
--- Returns all keys from label_name_map (for v2 APIs)
--- Handler code can filter by non-empty values for v1 legacy API support
+-- Returns underscored tag keys (for v2 APIs)
+-- Legacy API uses denormalizer to convert to dotted names
 SELECT DISTINCT key::text AS tag_key
 FROM metric_seg,
      LATERAL jsonb_object_keys(label_name_map) AS key
 WHERE organization_id = @organization_id
-  AND @metric_fingerprint = ANY(fingerprints)
+  AND @metric_fingerprint::BIGINT = ANY(fingerprints)
   AND label_name_map IS NOT NULL
 ORDER BY tag_key;
